Settle the two-level cell numbering at build finish instead of on the first lookup
`offsetInternal` no longer checks whether the prefix sums exist. The caller
computes them with `computeBucketPrefix` once the buckets stop growing, and
`HashJoin::onBuildPhaseFinish` does that for its two-level maps, after
`ConcurrentHashJoin` has merged the per-slot buckets into one map. That removes
an acquire load from every offset read on the probe path, and the table drops
the `std::once_flag` and the inline prefix array in favour of a vector that
exists only once it is computed.

`bits_for_bucket` is an `Int32`, and `bucketFor` stays public: it names the
sub-table of a key, which for fixed-range storage is not the bucket the key
routes to.

+++ src/Common/HashTable/HashSet.h
@@ -144,7 +144,7 @@ template <
     typename Hash = DefaultHash<Key>,
     typename Grower = TwoLevelHashTableGrower<>,
     typename Allocator = HashTableAllocator,
-    size_t bits_for_bucket = DEFAULT_BITS_FOR_BUCKET>
+    Int32 bits_for_bucket = DEFAULT_BITS_FOR_BUCKET>
 class TwoLevelHashSetTable
     : public TwoLevelHashTable<Key, TCell, Hash, Grower, Allocator, HashSetTable<Key, TCell, Hash, Grower, Allocator>, bits_for_bucket>
 {
@@ -232,7 +232,7 @@ template <
     typename Hash = DefaultHash<Key>,
     typename Grower = TwoLevelHashTableGrower<>,
     typename Allocator = HashTableAllocator,
-    size_t bits_for_bucket = DEFAULT_BITS_FOR_BUCKET>
+    Int32 bits_for_bucket = DEFAULT_BITS_FOR_BUCKET>
 using TwoLevelHashSet = TwoLevelHashSetTable<Key, HashTableCell<Key, Hash>, Hash, Grower, Allocator, bits_for_bucket>;
 
 template <typename Key, typename Hash, size_t initial_size_degree>
@@ -256,7 +256,7 @@ template <
     typename Hash = DefaultHash<Key>,
     typename Grower = TwoLevelHashTableGrower<>,
     typename Allocator = HashTableAllocator,
-    size_t bits_for_bucket = DEFAULT_BITS_FOR_BUCKET>
+    Int32 bits_for_bucket = DEFAULT_BITS_FOR_BUCKET>
 using TwoLevelHashSetWithSavedHash
     = TwoLevelHashSetTable<Key, HashSetCellWithSavedHash<Key, Hash>, Hash, Grower, Allocator, bits_for_bucket>;
